Added api request handling
Handling api requests is the same as handling actions. However
they are not event-based. So we need a way to trigger them. I used
update method of each ether for this purpose. I added new updateEnhancers
in order to add extra subscribtion for different kinds of observing.

useUpdateEnhancers is created for this purpose. And it uses Request enhancer
for listening state to see if we should update an API request. Requests use
a different kind of template then actions and state components. They also use
batching which means when we add different dependency list for each same fetch
component, it adds them together and listen for each state value change for this
deps.

We also have a new window property: getRequestTree for displaying request status
in the app state. injectEther now takes an options prop. It has only one child for
now: cacheSize. I implemented a cache logic for API requests in ether. If the same
request triggers more than once with same configuration then we use the value from the
cache instead of using fetch.

For api requests use requests prop in injectEther. each request is a method which returns
a fetch config object with url prop in it. Ether has also a short memory for api calls if
the current api call has the same configuration with the previous one it doesn't make a call
and update.